refs : #18066-Create Docker container for Elasticsearch with test data insertion
