+++ logharbour/test/testData/getLogs_testdata/1stTestcase.json
@@ -1,0 +1,82 @@
+[
+    {
+        "app": "crux",
+        "system": "merceadm-Latitude-3420",
+        "module": "",
+        "type": "A",
+        "pri": "Info",
+        "when": "2024-02-29T07:28:00.099424231Z",
+        "who": "Tushar",
+        "op": "",
+        "class": "wfinstance",
+        "instance": "",
+        "status": 0,
+        "remote_ip": "192.168.1.100",
+        "msg": "GetWFinstanceNew request received",
+        "data": "{}"
+    },
+    {
+        "app": "crux",
+        "system": "merceadm-Latitude-3420",
+        "module": "",
+        "type": "A",
+        "pri": "Info",
+        "when": "2024-02-29T07:28:00.099424231Z",
+        "who": "Tushar",
+        "op": "",
+        "class": "wfinstance",
+        "instance": "",
+        "status": 0,
+        "remote_ip": "192.168.1.100",
+        "msg": "GetWFinstanceNew request received",
+        "data": "{}"
+    },
+    {
+        "app": "crux",
+        "system": "merceadm-Latitude-3420",
+        "module": "",
+        "type": "A",
+        "pri": "Info",
+        "when": "2024-02-29T07:28:00.099424231Z",
+        "who": "Tushar",
+        "op": "",
+        "class": "wfinstance",
+        "instance": "",
+        "status": 0,
+        "remote_ip": "192.168.1.100",
+        "msg": "GetWFinstanceNew request received",
+        "data": "{}"
+    },
+    {
+        "app": "crux",
+        "system": "merceadm-Latitude-3420",
+        "module": "",
+        "type": "A",
+        "pri": "Info",
+        "when": "2024-02-25T07:28:00.10259862Z",
+        "who": "Tushar",
+        "op": "",
+        "class": "",
+        "instance": "",
+        "status": 0,
+        "remote_ip": "198.51.100.55",
+        "msg": "validating entity attributes",
+        "data": "{}"
+    },
+    {
+        "app": "crux",
+        "system": "merceadm-Latitude-3420",
+        "module": "",
+        "type": "A",
+        "pri": "Debug0",
+        "when": "2024-02-25T07:28:00.102604301Z",
+        "who": "Tushar",
+        "op": "",
+        "class": "wfinstance",
+        "instance": "",
+        "status": 0,
+        "remote_ip": "198.51.100.55",
+        "msg": "wfinstanceNewRequest after adding additional attributes :",
+        "data": "{\"slice\":2,\"app\":\"retailBANK\",\"entityid\":\"0eb8da50-aece-11ee-b168-3b192f7cd2b6\",\"entity_have_to_change\":{\"ageinstock\":\"2\",\"class\":\"inventoryitems\",\"fullname\":\"belampally\",\"inventoryqty\":\"2\",\"mrp\":\"200.00\",\"step\":\"START\",\"stepfailed\":\"false\"},\"workflow\":\"discountcheck\",\"trace\":0,\"parent\":77}"
+    }
+]
